requirements.txt
openai
requests
faiss-cpu
python-dotenv
mlflow
pytest
dotenv
langchain_openai;
